backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-pptx==0.6.21
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0